feat: Transactions CRUD and transactions CRUD tests

+++ fin_control/routers/users_router.py
@@ -13,6 +13,9 @@ router = APIRouter(prefix='/users', tags=['Users'])
 
 @router.get('/', status_code=HTTPStatus.OK, response_model=UserList)
 async def read_users(session: T_Session, request_user: Request_User, pagination: T_Filter_Page):
+    if not request_user.superuser:
+        raise HTTPException(status_code=HTTPStatus.UNAUTHORIZED, detail='Not Enough Permissions')
+
     offset = (pagination.page - 1) * pagination.limit
     users = await session.scalars(Select(User).limit(pagination.limit).offset(offset).order_by('id'))
 
@@ -25,7 +28,7 @@ async def get_user(user_id: int, session: T_Session, request_user: Request_User)
     if not user:
         raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail='User not found')
 
-    if user_id != request_user.id:
+    if not request_user.superuser and user_id != request_user.id:
         raise HTTPException(status_code=HTTPStatus.UNAUTHORIZED, detail='Not Enough Permissions')
 
     return user
@@ -57,11 +60,11 @@ async def create_user(user: UserSchema, session: T_Session):
 
 
 @router.patch('/{user_id}', status_code=HTTPStatus.OK, response_model=UserPublic)
-async def update_user(user_id: int, user: UserUpdate, session: T_Session):
-    user_object = await session.scalar(Select(User).where(User.id == user_id))
+async def update_user(user_id: int, user: UserUpdate, session: T_Session, request_user: Request_User):
+    if user_id != request_user.id:
+        raise HTTPException(status_code=HTTPStatus.UNAUTHORIZED, detail='Not Enough Permissions')
 
-    if not user_object:
-        raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail='User not found')
+    user_object = request_user
 
     if user.username:
         user_exists = await session.scalar(Select(User).where(User.username == user.username))
@@ -91,13 +94,16 @@ async def update_user(user_id: int, user: UserUpdate, session: T_Session):
 
 
 @router.delete('/{user_id}', status_code=HTTPStatus.OK)
-async def delete_user(user_id: int, session: T_Session):
+async def delete_user(user_id: int, session: T_Session, request_user: Request_User):
+    if not request_user.superuser:
+        raise HTTPException(status_code=HTTPStatus.UNAUTHORIZED, detail='Not Enough Permissions')
+
     user_object = await session.scalar(Select(User).where(User.id == user_id))
 
     if not user_object:
         raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail='User not found')
 
-    session.delete(user_object)
+    await session.delete(user_object)
     await session.commit()
 
     return {'message': 'Operation successfully'}
